+++ ui/src/app/core/settings.interfaces.ts
@@ -67,6 +67,8 @@ export interface AppSettingsInterface {
   env: EnvInterface
   formAuth: boolean
   host?: string
+  webroot?: string
+  originalWebroot?: string
   proxyHost?: string
   sessionTimeout: number
   sessionTimeoutInactivityBased: boolean